More tightening up the animations
Jumping animation occurs more instantly.

Fixed a bug where the play would stay in the jumping state on the ground if the jump button was held.

Added transition from watering to jumping.

Water particle object is now a child of a new object at the end of the staff to make the water appear more like it is coming out of the staff.

Sun is also now a child of the staff.

Commented out more debug statements.